move persistence and device id validation to core
